+++ changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 
+### v3.2.1 (04.12.2023)
+
+Self-filtering option for individual filter (disabled by default). [Feature Request](https://github.com/k-samuel/faceted-search/issues/37)
+Advanced configuration for AggregationQuery, if enabled, then result for filter can contain only filter values.
+Useful for cases with ValueIntersectionFilter  (AND condition).
+```php
+$filters[] = (new ValueIntersectionFilter('size', [12,32]))->selfFiltering(true);
+```
+
+
+
 ### v3.2.0 (29.11.2023)
 
 - ValueIntersectionFilter added [Feature Request](https://github.com/k-samuel/faceted-search/issues/33)